+++ app/src/main/java/com/example/myapplication/models/Album.kt
@@ -1,0 +1,11 @@
+package com.example.myapplication.models
+
+data class Album (
+    val albumId:Int,
+    val name:String,
+    val cover:String,
+    val releaseDate:String,
+    val description:String,
+    val genre:String,
+    val recordLabel:String
+)
