+++ app/src/main/java/com/example/myapplication/models/Collector.kt
@@ -1,0 +1,8 @@
+package com.example.myapplication.models
+
+data class Collector (
+    val collectorId: Int,
+    val name:String,
+    val telephone:String,
+    val email:String,
+)
