+++ app/src/main/java/com/example/myapplication/viewmodels/AlbumListViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.album
+package com.example.myapplication.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
