+++ app/src/main/res/layout/fragment_album_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.album.AlbumListFragment">
+    tools:context=".ui.AlbumListFragment">
 
     <TextView
         android:id="@+id/text_album_list"
